Added missing classes/args to match the HTTP APIs between browser and native
