+++ lib/browser/io.ts
@@ -41,6 +41,61 @@ export class InputStream {
 }
 
 /**
+ * Represents resources required to bootstrap a client connection, provided as
+ * a stub for the browser API
+ *
+ * @module aws-crt
+ * @category I/O
+ */
+export class ClientBootstrap { };
+
+/**
+ * Options for creating a {@link ClientTlsContext}. Provided as a stub for
+ * browser API.
+ *
+ * @module aws-crt
+ * @category TLS
+ */
+export type TlsContextOptions = any;
+
+/**
+ * TLS options that are unique to a given connection using a shared TlsContext.
+ * Provided as a stub for browser API.
+ *
+ * @module aws-crt
+ * @category TLS
+ */
+export class TlsConnectionOptions {
+    constructor(readonly tls_ctx: TlsContext, readonly server_name?: string, readonly alpn_list: string[] = []) {
+
+    }
+};
+
+/**
+ * TLS context used for TLS communications over sockets. Provided as a
+ * stub for the browser API
+ *
+ * @module aws-crt
+ * @category TLS
+ */
+export abstract class TlsContext {
+
+};
+
+/**
+ * TLS context used for client TLS communications over sockets. Provided as a
+ * stub for the browser API
+ *
+ * @module aws-crt
+ * @category TLS
+ */
+export class ClientTlsContext extends TlsContext {
+    constructor(options?: TlsContextOptions) {
+        super();
+    }
+};
+
+/**
  * Standard Berkeley socket style options.
  *
  * Provided for compatibility with nodejs, but this version is largely unused.
